Finalize navigation
