+++ cmd/app.go
@@ -1,22 +1,23 @@
 package cmd
 
 import (
-  "fmt"
-  "github.com/ESELDevelopment/ecmanager/internal/pages"
-  "os"
+	"fmt"
+	"github.com/ESELDevelopment/ecmanager/internal/pages"
+	"github.com/ESELDevelopment/ecmanager/internal/pages/start"
+	"os"
 
-  tea "github.com/charmbracelet/bubbletea"
+	tea "github.com/charmbracelet/bubbletea"
 )
 
 func Start() {
-  p := tea.NewProgram(
-    pages.Router(),
-    tea.WithAltScreen(),
-    tea.WithMouseAllMotion(),
-    tea.WithFPS(120),
-  )
-  if _, err := p.Run(); err != nil {
-    fmt.Println(err)
-    os.Exit(1)
-  }
+	p := tea.NewProgram(
+		start.New(pages.CreateRouter()),
+		tea.WithAltScreen(),
+		tea.WithMouseAllMotion(),
+		tea.WithFPS(120),
+	)
+	if _, err := p.Run(); err != nil {
+		fmt.Println(err)
+		os.Exit(1)
+	}
 }
